# openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  models.UserResponse:
    properties:
      address:
        type: string
      confirmed_at:
        type: string
      created_at:
        type: string
      credit_score:
        type: integer
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      identity:
        type: string
      locale:
        type: string
      phone:
        type: string
      role:
        type: string
      rtn:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  services.LoginResult:
    properties:
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Fintera Real Estate Management System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fintera API
  version: "1.0"
paths:
  /analytics/distribution:
    get:
      description: Returns availability statistics for lots
      parameters:
      - description: Project ID
        in: query
        name: project_id
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get Lot Distribution
      tags:
      - Analytics
  /analytics/export:
    get:
      description: Generates and downloads analytics reports in various formats
      parameters:
      - description: Report format (csv, xlsx, pdf)
        in: query
        name: format
        required: true
        type: string
      - description: Project ID
        in: query
        name: project_id
        type: integer
      - description: Start Date (ISO 8601)
        in: query
        name: start_date
        type: string
      - description: End Date (ISO 8601)
        in: query
        name: end_date
        type: string
      produces:
      - application/octet-stream
      responses: {}
      security:
      - BearerAuth: []
      summary: Export Analytics Data
      tags:
      - Analytics
  /analytics/overview:
    get:
      description: Returns high-level statistics and trend data
      parameters:
      - description: Project ID
        in: query
        name: project_id
        type: integer
      - description: Start Date (ISO 8601)
        in: query
        name: start_date
        type: string
      - description: End Date (ISO 8601)
        in: query
        name: end_date
        type: string
      - description: Revenue timeframe (6M or 12M)
        in: query
        name: revenue_timeframe
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get Analytics Overview
      tags:
      - Analytics
  /analytics/performance:
    get:
      description: Returns performance metrics for multiple projects
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get Project Performance
      tags:
      - Analytics
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.LoginResult'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logs out a user (invalidates refresh token)
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.LoginResult'
      summary: Refresh Token
      tags:
      - Auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"